+++ src/command/DevMode.ts
@@ -1,3 +1,5 @@
+import * as vscode from 'vscode';
+import { statusBarItem } from '../configUI';
 import type { TAhkFileData } from '../core/ProjectManager';
 import { OutputChannel } from '../provider/vscWindows/OutputChannel';
 import {
@@ -12,12 +14,16 @@ import { UpdateCacheAsync } from './UpdateCache';
 
 const DevModeData: number[] = [];
 
-function devTestDA(): void {
+function devTestDA(cycles: number): void {
     const t1: number = Date.now();
 
     void UpdateCacheAsync(true)
         .then((ed: TAhkFileData[] | null): null => {
-            if (ed !== null) DevModeData.push(Date.now() - t1);
+            if (ed !== null) {
+                DevModeData.push(Date.now() - t1);
+                statusBarItem.text = `$(heart) ${DevModeData.length} of ${cycles}`;
+                statusBarItem.show();
+            }
             return null;
         })
         .catch((error: Error): void => {
@@ -25,7 +31,7 @@ function devTestDA(): void {
         });
 }
 
-function devTestEnd(iMax: number): void {
+function devTestEnd(cycles: number): void {
     const statistics: number[] = [...DevModeData];
     DevModeData.length = 0;
 
@@ -39,8 +45,7 @@ function devTestEnd(iMax: number): void {
     OutputChannel.appendLine([
         '---------------------------------------------',
         'The task be completed, please confirm!',
-        `iMax is ${iMax}`,
-        `statistics len is ${len}`,
+        `resolve ${len} of ${cycles}`,
         `sum is ${sum}`,
         `avg is ${avg}`,
         `stdDev is ${stdDev}`,
@@ -54,7 +59,20 @@ function devTestEnd(iMax: number): void {
         '---------------------------------------------',
     ].join('\n'));
 
+    statusBarItem.text = '$(heart) dev task OK!';
+    statusBarItem.show();
+
     OutputChannel.show();
+}
+
+function validateInput(value: string): vscode.InputBoxValidationMessage | null {
+    if ((/^[1-9]\d+/u).test(value)) {
+        return null;
+    }
+    return {
+        message: 'should input ms of > 10 integer, like 80 or 90',
+        severity: vscode.InputBoxValidationSeverity.Error,
+    };
 }
 
 const TimeoutList: NodeJS.Timeout[] = [];
@@ -69,38 +87,41 @@ export async function pressureTest(): Promise<null> {
     const pick: TPickReturn | undefined = await pressureTestConfig();
     if (pick === undefined) return null;
 
-    const {
-        maxTime,
-        delay,
-        label,
-    } = pick;
+    const msStr: string | undefined = await vscode.window.showInputBox({
+        title: 'input delay ms of a cycles',
+        prompt: 'Recommended reference of `Refresh Resource` * 1.5 ~ 2',
+        placeHolder: '150',
+        validateInput,
+    });
+    if (msStr === undefined) return null;
+
+    const { cycles, label } = pick;
 
     OutputChannel.appendLine('---------------------------------------------');
     OutputChannel.appendLine('>> this is Dev tools, open "vscode-js-profile-flame" to get ".cpuprofile"');
     OutputChannel.appendLine(`    please wait of [${label}]`);
     OutputChannel.show();
 
-    for (let i = 1; i <= maxTime; i++) {
-        TimeoutList.push(setTimeout(devTestDA, i * delay));
+    const delay: number = Number.parseInt(msStr, 10);
+    for (let i = 1; i <= cycles; i++) {
+        TimeoutList.push(setTimeout(devTestDA, i * delay, cycles));
     }
 
     // eslint-disable-next-line no-magic-numbers
-    TimeoutList.push(setTimeout(devTestEnd, (maxTime + 5) * delay, maxTime));
+    TimeoutList.push(setTimeout(devTestEnd, (cycles + 5) * delay, cycles));
     return null;
 }
 
-// ---------------------------------------------
 // The task be completed, please confirm!
-// iMax is 80
-// statistics len is 80
-// sum is 13154
-// avg is 164.425
-// stdDev is 5.2211231150278845
-// []
+// resolve 80 of 80
+// sum is 6573
+// avg is 82.1625
+// stdDev is 5.775206172000611
+// [128, 95, 89, 85, 85, 85, 86, 84, 83, 83, 83, 82, 82, 84, ...]
 // ---Min avg of 5 ---
-// subAvg is 166.8
-// subAvgArr len is [168, 166, 167, 166, 167]
+// subAvg is 79.2
+// subAvgArr len is [80, 79, 79, 79, 79]
 // ---Min std of 5 ---
-// subStd is 0
-// subStdArr len is [168, 168, 168, 168, 168]
+// subStd is 0.4472135954999579
+// subStdArr len is [80, 81, 80, 80, 80]
 // ---------------------------------------------